# panel/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Marketing Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspaces",
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceOut"
                  },
                  "type": "array",
                  "title": "Response List Workspaces Api V1 Workspaces Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace",
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "workspaces"
        ],
        "summary": "Update Workspace",
        "operationId": "update_workspace_api_v1_workspaces__workspace_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/calendar": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Calendar",
        "description": "Every scheduled/published post plus known holidays in one range\n(handoff section 11) — what backs the panel's drag-and-drop calendar.",
        "operationId": "get_calendar_api_v1_workspaces__workspace_id__calendar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/briefs": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Brief",
        "description": "Store a new brief version.\n\nBriefs are never edited in place — the agents that produced a package must\nstay traceable to the exact brief they read.",
        "operationId": "create_brief_api_v1_workspaces__workspace_id__briefs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandBriefCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandBriefOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Briefs",
        "operationId": "list_briefs_api_v1_workspaces__workspace_id__briefs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrandBriefOut"
                  },
                  "title": "Response List Briefs Api V1 Workspaces  Workspace Id  Briefs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/briefs/active": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Active Brief",
        "operationId": "active_brief_api_v1_workspaces__workspace_id__briefs_active_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandBriefOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandBriefCreate": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BrandBriefData"
          },
          "activate": {
            "type": "boolean",
            "title": "Activate",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BrandBriefCreate"
      },
      "BrandBriefData": {
        "properties": {
          "brand": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Brand"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          },
          "website": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "voice": {
            "$ref": "#/components/schemas/BrandVoice"
          },
          "visual_identity": {
            "$ref": "#/components/schemas/VisualIdentity"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/Competitor"
            },
            "type": "array",
            "title": "Competitors"
          },
          "personas": {
            "items": {
              "$ref": "#/components/schemas/Persona"
            },
            "type": "array",
            "title": "Personas"
          },
          "seed_keywords": {
            "$ref": "#/components/schemas/SeedKeywords"
          },
          "pillars": {
            "items": {
              "$ref": "#/components/schemas/Pillar"
            },
            "type": "array",
            "title": "Pillars"
          },
          "ai_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ai Questions"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "type": "array",
            "title": "Channels"
          },
          "goals": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Goals"
          },
          "video_mode": {
            "$ref": "#/components/schemas/VideoMode",
            "default": "voice"
          },
          "locales": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locales"
          },
          "evidence_policy": {
            "type": "string",
            "title": "Evidence Policy",
            "default": "every statistic must carry a source URL"
          }
        },
        "type": "object",
        "required": [
          "brand"
        ],
        "title": "BrandBriefData"
      },
      "BrandBriefOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "video_mode": {
            "$ref": "#/components/schemas/VideoMode"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "version",
          "is_active",
          "video_mode",
          "data",
          "created_at"
        ],
        "title": "BrandBriefOut"
      },
      "BrandVoice": {
        "properties": {
          "tone": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tone"
          },
          "person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person",
            "description": "e.g. first plural, second singular"
          },
          "do": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Do"
          },
          "dont": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dont"
          },
          "reading_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Level"
          }
        },
        "type": "object",
        "title": "BrandVoice"
      },
      "CalendarHolidayOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "name_fa": {
            "type": "string",
            "title": "Name Fa"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          }
        },
        "type": "object",
        "required": [
          "date",
          "name_fa",
          "name_en"
        ],
        "title": "CalendarHolidayOut"
      },
      "CalendarOut": {
        "properties": {
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "calendar": {
            "type": "string",
            "title": "Calendar"
          },
          "min_publish_spacing_minutes": {
            "type": "integer",
            "title": "Min Publish Spacing Minutes"
          },
          "publications": {
            "items": {
              "$ref": "#/components/schemas/CalendarPublicationOut"
            },
            "type": "array",
            "title": "Publications"
          },
          "holidays": {
            "items": {
              "$ref": "#/components/schemas/CalendarHolidayOut"
            },
            "type": "array",
            "title": "Holidays"
          }
        },
        "type": "object",
        "required": [
          "timezone",
          "calendar",
          "min_publish_spacing_minutes",
          "publications",
          "holidays"
        ],
        "title": "CalendarOut"
      },
      "CalendarPublicationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "package_id": {
            "type": "string",
            "format": "uuid",
            "title": "Package Id"
          },
          "package_title": {
            "type": "string",
            "title": "Package Title"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "package_id",
          "package_title",
          "channel",
          "status",
          "scheduled_at",
          "external_url"
        ],
        "title": "CalendarPublicationOut"
      },
      "Channel": {
        "type": "string",
        "enum": [
          "wordpress",
          "telegram",
          "instagram",
          "linkedin",
          "x",
          "youtube",
          "aparat"
        ],
        "title": "Channel",
        "description": "Handoff section 10."
      },
...
      "Competitor": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Competitor"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicationStatus": {
        "type": "string",
        "enum": [
          "scheduled",
          "publishing",
          "published",
          "failed",
          "cancelled"
        ],
        "title": "PublicationStatus"
      },
...
      "SeedKeywords": {
        "properties": {
          "fa": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fa"
          },
          "en": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "En"
          },
          "ar": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ar"
          }
        },
        "type": "object",
        "title": "SeedKeywords",
        "description": "Keywords are researched per language, never translated (decision D3)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoMode": {
        "type": "string",
        "enum": [
          "none",
          "voice",
          "face"
        ],
        "title": "VideoMode",
        "description": "Decision D4."
      },
      "VisualIdentity": {
        "properties": {
          "palette": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Palette"
          },
          "fonts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fonts"
          },
          "style_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Style Keywords"
          },
          "logo_storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Storage Key"
          },
          "overlay_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Template"
          }
        },
        "type": "object",
        "title": "VisualIdentity"
      },
...
      "WorkspaceCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]([a-z0-9-]{0,62}[a-z0-9])?$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "default_locale": {
            "type": "string",
            "title": "Default Locale",
            "default": "fa"
          },
          "locales": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locales"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Tehran"
          },
          "calendar": {
            "type": "string",
            "title": "Calendar",
            "default": "jalali"
          },
          "min_publish_spacing_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Min Publish Spacing Minutes",
            "default": 60
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "WorkspaceCreate"
      },
      "WorkspaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "default_locale": {
            "type": "string",
            "title": "Default Locale"
          },
          "locales": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locales"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "calendar": {
            "type": "string",
            "title": "Calendar"
          },
          "min_publish_spacing_minutes": {
            "type": "integer",
            "title": "Min Publish Spacing Minutes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "slug",
          "name",
          "default_locale",
          "locales",
          "timezone",
          "calendar",
          "min_publish_spacing_minutes",
          "is_active",
          "created_at"
        ],
        "title": "WorkspaceOut"
      },
      "WorkspaceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "default_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Locale"
          },
          "locales": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locales"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "calendar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar"
          },
          "min_publish_spacing_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Publish Spacing Minutes"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "WorkspaceUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}